Add module for single task scheduling and execution
Similar to Unix cron the leader_cron_task module schedules a periodic
task to be executed repeatedly in the future. The schedule is defined
by five fields

Field         Valid Range
minute        0 - 59
hour          0 - 23
day of month  1 - 31
month         1 - 12
day of week   0 - 6 (Sunday is 0)

The schedule is defined by the cron tuple

{cron, {Minute, Hour, DayOfMonth, Month, DayOfWeek}}

The semantics of these fields align with Unix cron. Each field
specifies which values in the range are valid for task execution. The
values can be given as a single value, a range, a list or the atom 'all'.

Field Spec                     Example            Unix Cron
all                            all                *
{range, integer(), integer()}  {range, 1, 5}      1-5
{list, [integer()]}            {list, [1, 3, 7]}  1,3,7

If the day of month is set to a day which does not exist in the current
month (such as 31 for February) the day is skipped. Setting day of month
to 31 does _not_ mean the last day of the month. This aligns with Unix
cron.

Specified dates and times are all handled in UTC.

When a task takes longer than the time to the next valid period (or
periods) the overlapped periods are skipped.